+++ matflow/models/workflow.py
@@ -13,10 +13,7 @@ import shutil
 from datetime import datetime
 from enum import Enum
 from pathlib import Path
-from pprint import pprint
-from subprocess import run, PIPE
 from warnings import warn
-import time
 
 import h5py
 import hickle
@@ -27,8 +24,6 @@ from ruamel.yaml import YAML, scalarstring
 from matflow import __version__
 from matflow.config import Config
 from matflow.errors import (
-    IncompatibleTaskNesting,
-    MissingMergePriority,
     WorkflowPersistenceError,
     TaskElementExecutionError,
     UnexpectedSourceMapReturnError,
@@ -49,7 +44,7 @@ from matflow.utils import (
 
 
 def requires_ids(func):
-    'Workflow method decorator to raise if IDs are not assigned.'
+    """Workflow method decorator to raise if IDs are not assigned."""
     @functools.wraps(func)
     def func_wrap(self, *args, **kwargs):
         if not self.id:
@@ -59,7 +54,7 @@ def requires_ids(func):
 
 
 def requires_path_exists(func):
-    'Workflow method decorator to raise if workflow path does not exist as a directory.'
+    """Workflow method decorator to raise if workflow path does not exist as a directory."""
     @functools.wraps(func)
     def func_wrap(self, *args, **kwargs):
         if not self.path_exists:
@@ -93,11 +88,12 @@ class Workflow(object):
         '_archive',
         '_archive_excludes',
         '_figures',
+        '_metadata',
     ]
 
     def __init__(self, name, tasks, stage_directory=None, extends=None, archive=None,
-                 archive_excludes=None, figures=None, check_integrity=True, profile=None,
-                 __is_from_file=False):
+                 archive_excludes=None, figures=None, metadata=None,
+                 check_integrity=True, profile=None, __is_from_file=False):
 
         self._id = None             # Assigned once by set_ids()
         self._human_id = None       # Assigned once by set_ids()
@@ -113,13 +109,14 @@ class Workflow(object):
         self._figures = [{'idx': idx, **i}
                          for idx, i in enumerate(figures)
                          ] if figures else []
+        self._metadata = metadata or {}
 
         tasks, elements_idx = init_tasks(self, tasks, self.is_from_file, check_integrity)
         self._tasks = tasks
         self._elements_idx = elements_idx
 
         if not self.is_from_file:
-            self._check_cloud_archive()
+            self._check_archive_connection()
             self._history = []
             self._append_history(WorkflowAction.generate)
 
@@ -160,13 +157,22 @@ class Workflow(object):
     def __len__(self):
         return len(self.tasks)
 
-    def _check_cloud_archive(self):
-        if self.archive and 'cloud_provider' in self.archive_definition:
+    def _check_archive_connection(self):
+        if not self.archive:
+            return
+
+        # TODO: should this whole function be outsourced to hpcflow?
+
+        # TODO: check local archive path exists
+
+        if 'cloud_provider' in self.archive_definition:
             provider = self.archive_definition['cloud_provider']
             hpcflow.cloud_connect(provider, config_dir=Config.get('hpcflow_config_dir'))
 
+        # TODO when supported in DataLight, add datalight.check_access(...)
+
     def _append_history(self, action, **kwargs):
-        'Append a new history event.'
+        """Append a new history event."""
 
         if action not in WorkflowAction:
             raise TypeError('`action` must be a `WorkflowAction`.')
@@ -238,12 +244,12 @@ class Workflow(object):
 
     @property
     def name_safe(self):
-        'Get name without spaces'
+        """Get name without spaces"""
         return self.name.replace(' ', '_')
 
     @property
     def name_friendly(self):
-        'Capitalise and remove underscores'
+        """Capitalise and remove underscores"""
         name = '{}{}'.format(self.name[0].upper(), self.name[1:]).replace('_', ' ')
         return name
 
@@ -262,6 +268,10 @@ class Workflow(object):
     @property
     def figures(self):
         return self._figures
+
+    @property
+    def metadata(self):
+        return self._metadata
 
     @property
     def elements_idx(self):
@@ -300,7 +310,7 @@ class Workflow(object):
 
     @property
     def path_exists(self):
-        'Does the Workflow project directory exist on this machine?'
+        """Does the Workflow project directory exist on this machine?"""
         try:
             path = self.path
         except ValueError:
@@ -313,12 +323,12 @@ class Workflow(object):
     @property
     @requires_ids
     def path(self):
-        'Get the full path of the Workflow project as a Path.'
+        """Get the full path of the Workflow project as a Path."""
         return Path(self.stage_directory, self.human_id)
 
     @property
     def path_str(self):
-        'Get the full path of the Workflow project as a string.'
+        """Get the full path of the Workflow project as a string."""
         return str(self.path)
 
     @property
@@ -357,7 +367,7 @@ class Workflow(object):
             return out
 
     def get_task_idx_padded(self, task_idx, ret_zero_based=True):
-        'Get a task index, zero-padded according to the number of tasks.'
+        """Get a task index, zero-padded according to the number of tasks."""
         if ret_zero_based:
             return zeropad(task_idx, len(self) - 1)
         else:
@@ -365,7 +375,7 @@ class Workflow(object):
 
     @requires_path_exists
     def get_task_path(self, task_idx):
-        'Get the path to a task directory.'
+        """Get the path to a task directory."""
         if task_idx > (len(self) - 1):
             msg = f'Workflow has only {len(self)} tasks.'
             raise ValueError(msg)
@@ -381,7 +391,7 @@ class Workflow(object):
 
     @requires_path_exists
     def get_element_path(self, task_idx, element_idx):
-        'Get the path to an element directory.'
+        """Get the path to an element directory."""
         num_elements = self.elements_idx[task_idx]['num_elements']
         if element_idx > (num_elements - 1):
             msg = f'Task at index {task_idx} has only {num_elements} elements.'
@@ -416,7 +426,7 @@ class Workflow(object):
         return out
 
     def write_directories(self):
-        'Generate task and element directories.'
+        """Generate task and element directories."""
 
         if self.path.exists():
             raise ValueError('Directories for this workflow already exist.')
@@ -512,7 +522,7 @@ class Workflow(object):
 
     @requires_path_exists
     def get_hpcflow_workflow(self):
-        'Generate an hpcflow workflow to execute this workflow.'
+        """Generate an hpcflow workflow to execute this workflow."""
 
         command_groups = []
         variables = {}
@@ -846,7 +856,7 @@ class Workflow(object):
             return None
 
     def as_dict(self):
-        'Return attributes dict with preceding underscores removed.'
+        """Return attributes dict with preceding underscores removed."""
         out = {k.lstrip('_'): getattr(self, k) for k in self.__slots__}
 
         # Deal with the WorkflowAction enums and datetimes in history action values:
@@ -1114,11 +1124,21 @@ class Workflow(object):
         obj = {
             'name': obj_json['name'],
             'tasks': obj_json['tasks'],
-            'figures': obj_json['figures'],
             'stage_directory': obj_json['stage_directory'],
             'profile': obj_json['profile'],
             'extends': obj_json['extends'],
         }
+
+        # For loading older workflow files without these attributes:
+        WARN_ON_MISSING = [
+            'figures',
+            'metadata',
+        ]
+        for key in WARN_ON_MISSING:
+            if key not in obj_json:
+                warn(f'"{key}" key missing from this workflow.')
+            else:
+                obj.update({key: obj_json[key]})
 
         workflow = cls(
             **obj,
@@ -1223,7 +1243,7 @@ class Workflow(object):
 
     @requires_path_exists
     def prepare_sources(self, task_idx):
-        'Prepare source files for the task preparation commands.'
+        """Prepare source files for the task preparation commands."""
 
         # Note: in future, we might want to parametrise the source function, which is
         # why we delay its invocation until task run time.
@@ -1325,7 +1345,7 @@ class Workflow(object):
 
     @requires_path_exists
     def run_python_task(self, task_idx, element_idx):
-        'Execute a task that is to be run directly in Python (via the function mapper).'
+        """Execute a task that is to be run directly in Python (via the function mapper)."""
 
         task = self.tasks[task_idx]
         element = task.elements[element_idx]